// types/types.d.ts
export interface User extends DocumentNode, basicTable {
  password?: string
  photo?: fileType
  name?: string
  lastname?: string
  email?: string
  privilegeID?: Privilege
  active?: boolean
  tokenExpo?: string
  token?: string
  isActive?: string
  rolName?: string
  canCreateHost: boolean
  allEventWithAuth: boolean
  verifyLogin?: boolean
  lang?: string
  canAccessToApp?: boolean
  canAccessToWeb: ?boolean
  canUseAuthenticator?: boolean
  document: string
  typeDocument: string
  code?: boolean
  phone?: string
  QR?: string
  termporal_Qr?: IWorker_qr_temporal
  group: IGroupWorker[]
  banFinish: string
}
// ...
export type Privilege = {
  _id?: string
  key?: string
  name: privilegeName
  createdAt?: Date
  UpdatedAt?: Date
  permissions?: PermissionsPrivilege[]
}
// ...
export interface IContact extends Document, basicTable {
  firstName?: string
  lastName?: string
  email?: string
  DPI: string
  phone?: string
  indicativo?: string
  host?: User | string
  nickname?: string
  verified?: boolean
  askVerification?: boolean | string
  verifiedData?: verifiedData
  verifiedDataPDF?: verifiedDataPDF
  banFinish?: string
  ban?: boolean
  disabled?: boolean
}

export type verifiedDataPDF = {
  photo?: fileType
  documentA?: fileType
  documentB?: fileType
  num1?: string
  type?: string
  name?: string
  expedition?: string
  expiration?: string
  licNum?: string
  num2?: string
}

export type ReadedMRZ = {
  birthDate?: string
  expirationDate?: string
  sex?: string
  firstName?: string
  lastName?: string
  nationality?: string
  documentNumber?: string
}

export type ReadedPDF = {
  num1?: string
  type?: string
  name?: string
  expedition?: string
  expiration?: string
  licNum?: string
  num2?: string
}

export type fileType = {
  filename: string
  key: string
}

export type verifiedData = {
  photo?: fileType
  documentA?: fileType
  documentB?: fileType
  birthDate?: string
  expirationDate?: string
  sex?: string
  lastName?: string
  firstName?: string
  nationality?: string
  documentNumber?: string
}
// ...
export interface IGroupWorker extends Document, basicTable {
  name: string
  locations: ILocation[]
  abbreviation: string
  exists: boolean
}
// ...
export interface IWorker_qr_temporal extends Document, basicTable {
  QR: string
  timeEnd: string
  used: boolean
  valid: boolean
  location: ILocation
}